Agregacion de Controllers

+++ core/src/main/java/com/undec/AppClima/output/IModifyUserRepository.java
@@ -3,6 +3,6 @@ import com.undec.AppClima.domain.User;
 public interface IModifyUserRepository {
     boolean userMailExist(String mail);
     boolean userNameExist(String userName);
-    User get_Current_User(int idCurrentUser);
-    boolean modifyUser(User user,int idCurrentUser);
+    User get_Current_User(Long idCurrentUser);
+    boolean modifyUser(User user,Long idCurrentUser);
 }
